requirements.txt
flask==3.0.0
requests==2.31.0
spacy==3.7.2
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
lyricsgenius==3.0.1
python-dotenv==1.0.0

